fix(channelPage): 채널 페이지 팔로워 수 오류 수정 - ChannelProfile 컴포넌트 팔로워 상태 관리 로직 제거 - 팔로우 상태 관리 로직 추가
