+++ src/app/(route)/channel/[uid]/layout.tsx
@@ -8,6 +8,7 @@ import useNavToggle from "@/app/_store/main/useNavToggle.client";
 import { createClient } from "../../../_utils/supabase/client";
 import ChannelProfile from "@/app/(route)/channel/[uid]/components/ChannelProfile";
 import Footer from '@/app/_components/Footer/footer';
+import { useFollowAction } from '@/app/_store/queries/follow/mutation';
 
 interface RootLayoutProps {
   children: React.ReactNode;
@@ -17,14 +18,20 @@ const RootLayout = ({ children }: RootLayoutProps) => {
   const { isOpen } = useNavToggle();
   const [isClient, setIsClient] = useState(false); 
   const [userInfo, setUserInfo] = useState<{ nickname: string; channel_intro: string; img_url: string } | null>(null);
+  const [followerCount, setFollowerCount] = useState<number>(0);
+  const [isFollowing, setIsFollowing] = useState<boolean>(false);
+  const [loggedInUserId, setLoggedInUserId] = useState<string | null>(null);
+
   const { uid } = useParams(); 
   const router = useRouter();
   const activeSegment = useSelectedLayoutSegment();
   const isCommunityActive = !activeSegment || ['post', 'comment', 'detail', 'edit'].includes(activeSegment || '');
 
+  const supabase = createClient();
+  const { followMutate, unfollowMutate } = useFollowAction();
+
   const fetchUserInfo = async () => {
     if (!uid) return; 
-    const supabase = createClient();
     const { data, error } = await supabase
       .from("users")
       .select("nickname, channel_intro, profile_img")
@@ -42,8 +49,48 @@ const RootLayout = ({ children }: RootLayoutProps) => {
     }
   };
 
-  const handleTabClick = (path: string) => {
-    router.push(path);
+  const fetchLoggedInUser = async () => {
+    const {
+      data: { user },
+      error,
+    } = await supabase.auth.getUser();
+
+    if (error) {
+      console.error('로그인된 사용자 정보 가져오기 오류', error);
+    } else {
+      setLoggedInUserId(user?.id || null);
+    }
+  };
+
+  const fetchFollowerData = async () => {
+    if (!uid) return;
+    try {
+      const { count, error: countError } = await supabase
+        .from('follows')
+        .select('*', { count: 'exact' })
+        .eq('following_id', uid);
+
+      if (countError) {
+        throw new Error('팔로워 수 가져오기 오류: ' + countError.message);
+      }
+      setFollowerCount(count || 0);
+
+      if (loggedInUserId) {
+        const { data: followData, error: followError } = await supabase
+          .from('follows')
+          .select('*')
+          .eq('follower_id', loggedInUserId)
+          .eq('following_id', uid)
+          .single();
+
+        if (followError && followError.code !== 'PGRST116') {
+          throw new Error('팔로우 상태 확인 오류: ' + followError.message);
+        }
+        setIsFollowing(!!followData);
+      }
+    } catch (error) {
+      console.error(error);
+    }
   };
 
   useEffect(() => {
@@ -56,7 +103,63 @@ const RootLayout = ({ children }: RootLayoutProps) => {
     }
   }, [isClient, uid]);
 
+  useEffect(() => {
+    fetchLoggedInUser();
+  }, []);
+
+  useEffect(() => {
+    if (uid) {
+      fetchFollowerData();
+    }
+  }, [uid, loggedInUserId]);
+
+  const handleTabClick = (path: string) => {
+    router.push(path);
+  };
+
+  const handleChannelStudioClick = () => {
+    if (uid) router.push(`/studio/${uid}`);
+  };
+
+  const handleChannelManagementClick = () => {
+    if (uid) router.push(`/channel/${uid}/settings`);
+  };
+
+  const handleFollow = () => {
+    if (!uid) return;
+    followMutate(
+      { uid: Array.isArray(uid) ? uid[0] : uid, nickname: userInfo?.nickname || '' },
+      {
+        onSuccess: () => {
+          setIsFollowing(true);
+          setFollowerCount((prev) => prev + 1);
+        },
+        onError: (error) => {
+          console.error('팔로우 오류:', error);
+        },
+      }
+    );
+  };
+
+  const handleUnfollow = () => {
+    if (!uid) return;
+    unfollowMutate(
+      { uid: Array.isArray(uid) ? uid[0] : uid, nickname: userInfo?.nickname || '' }, 
+      {
+        onSuccess: () => {
+          setIsFollowing(false);
+          setFollowerCount((prev) => Math.max(prev - 1, 0));
+        },
+        onError: (error) => {
+          console.error('언팔로우 오류:', error);
+        },
+      }
+    );
+  };
+
   if (!userInfo) return <p>로딩 중...</p>;
+
+  const isLoggedInUser = loggedInUserId === uid;
 
   return (
     <div>
@@ -67,9 +170,14 @@ const RootLayout = ({ children }: RootLayoutProps) => {
         <ChannelProfile
           img_url={userInfo.img_url}
           nickname={userInfo.nickname}
-          follower={2.4} 
+          followerCount={followerCount}
           context={userInfo.channel_intro}
-          is_following={false}
+          isFollowing={isFollowing}
+          isLoggedInUser={isLoggedInUser}
+          onFollow={handleFollow}
+          onUnfollow={handleUnfollow}
+          onChannelStudioClick={handleChannelStudioClick}
+          onChannelManagementClick={handleChannelManagementClick}
         />
         <div className="flex flex-row mb-6">
           <p
